Prepare basic project files and their layout
Work on the project continued locally for a short time. The basis of the
project was taken from an earlier project:
https://github.com/Notiooo/Worms-Similar-Game.

The project contains basic code related to the game loop, calculation of
deltaTime as well as StateStack. Some OpenGL calls have been abstracted and the
camera and control have already been written before the repository was made.