Add missing libraries and fix the gitignore
The earlier gitignore contained the wrong setting of folders by which it
excluded the folder inside /vendors/Release/x64. Because of this, one of the
libraries did not appear on the repository.

This has now been fixed and the gitignore slightly changed so that this does
not occur in the future again.