Add the ability to highlight and delete blocks
ChunkMeshBuilder has been generalized to be a general MeshBuilder. This was
because there was nothing in it that limited its use to only chunks.

For variables like "BLOCKS_PER_DIMENSION" they were changed from unsigned to
int. This was due to conversion issues. At this point it has been resolved this
way.

Block removal now checks the blocks on the edge of the chunk to possibly
rebuild the chunk adjacent to the block.

The highlighting of the block is handled by a separate class "FacedBlock"