Add a basic chunk container with face-culling support
There was needed a Chunk Container base that would not only store and create
chunks, but most importantly create a valid mesh without walls between chunks.
For this purpose, the Chunk class was given the appropriate functions to
communicate between the container and check the blocks in the chunks next to
it.

The distinction between block coordinates and coordinates in space has been
properly marked.

For the future, support for easy resizing of blocks in space has been added.

In debug mode there is an option to change the texture pack.

Settings class has been slightly improved to not open and close the file too
often.

Class BlockMap was introduced, which allows to get information about existing
blocks and their properties. Blocks are stored in the form of files.

Handling of mouse behavior for first person mode has been moved to the mouse
class.