Add an infinite generation of the world
For this purpose, chunks are stored as std::shared_ptr<Chunk>. This is because
they can be deleted while still being rebuilt.

The building of chunk had to be moved to a separate thread, otherwise it would
cause momentary lags. Chunks that interact with the player have to be
prioritized here, because the player can't wait for the rest of the chunks to
generate because it would lead to a big lag. For this reason, this has been
handled with a queue.