Migrate to CMake and add possibilty to write tests
The CMake project automatically downloads the necessary projects and
then manually builds them and links them into the project. It also
allows to write tests to code and run them as a separate project. In
addition, smaller improvements have been made, such as changing some
folders to lowercase and adding a lock, without which the programme
could unexpectedly crash.