Improve the meaning of update and fixedUpdate
Previously update was a function that was called regardless of framerate
and fixedUpdate was called with deltaTime and was to be used when
something had to do with the number of frames produced. This behavior
has now changed to make sense by removing the low and fixed frame rate
threshold while bumping up performance.

FixedUpdate from now on has a certain minimum number of executions
within a certain amount of time - although it can be more. This is to
serve solutions such as collision checking, for example.

Update from now on is used both when the number of frames is relevant -
providing deltaTime for this purpose, but at the same time it can be
used by functions for which it is irrelevant.