Add simple main menu, and world saving/loading system
The menu is still unstyled, but writing the game state to a file and
reading it later works as well as possible. I wanted to use zpp_bits,
but at the time of writing the code, MSVC has a bug that doesn't allow
compiling zpp_bits, so I used the older serializer.h library.