Adds different biomes, and terrain smoothing between biomes of different heights
New biomes are determined by the noise corresponding to temperature and
humidity. Terrain is smoothed only on chunks that contain more than a
one biome including a chunk outline of 1 block around the chunk. In this
way, strange formations are not created in which two chunks with
completely different biomes collide side by side at the same time
without containing other biomes in them. In that case, you could see the
height differences. Checking the blocks around the chunk and
recalculating the average value on the four blocks in the chunk's corner
gives much better smoothing effects in such an edge case.